requirements.txt
python-docx
beautifulsoup4
Jinja2
PyDocX